Add parameter `supportDualStack` in Host creation
After testing we saw that the creation of the `Host` was crashing because
the `FleckWebSocketTransport` cannot create the web socket if we
cannot support the dual stack.

In a first time we believe that we could call the `RegisterTransport`
after the creation of the `Host` but an exception was thrown when we
create the object. So we need to give directly the `supportDualStack`
information in the `Host` creation.